move token api to 'api' application
