+++ project/backend/api/serializers.py
@@ -1,0 +1,5 @@
+from rest_framework import serializers
+
+class OauthCodeSerializer(serializers.Serializer):
+    code = serializers.CharField()
+    state = serializers.CharField()
